Move FragmenterResource to FragmenterService
fragmentIndex should start at 0
